Updated Controller Routes
Fixed Delete function in both Journal and Workout
TODO:  Separate Add and Edit into separate functions in both Journal and Workout

+++ PersonalNote/Models/JournalDataAccessLayer.cs
@@ -81,9 +81,12 @@ namespace PersonalNote.Models
         {
             try
             {
-                JournalReport journal = db.JournalReport.Find(id);
-                db.JournalReport.Remove(journal);
-                db.SaveChanges();
+                if (id > 0)
+                {
+                    JournalReport journal = db.JournalReport.Find(id);
+                    db.JournalReport.Remove(journal);
+                    db.SaveChanges();
+                }
             }
             catch
             {
